third_party/WebKit/LayoutTests/fast/forms/state-save-of-detached-control-expected.txt

Test to NOT save state of a detached form control.

PASS successfullyParsed is true

TEST COMPLETE
PASS document.getElementById("input2").value is ""
